pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minimalist_compliance_control"
version = "0.1.0"
description = "Minimal compliance control."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Songbo Hu"},
  {name = "Haochen Shi"}
]
dependencies = [
  "numpy>=1.23",
  "scipy>=1.9",
  "mujoco>=2.3.7",
  "mink>=0.0.6",
  "gin-config>=0.5.0",
]

[project.optional-dependencies]
examples = [
  "opencv-python",
  "pyyaml",
  "torch",
  "torchvision",
  "diffusers",
  "joblib",
  "pyzmq",
  "requests",
  "open3d",
  "pycocotools",
  "qpsolvers[osqp]",
  "sympy",
]

[tool.setuptools]
packages = [
  "minimalist_compliance_control",
  "minimalist_compliance_control.reference",
]
